Change how Bitbucket auth variables are shown
While following the getting started docs for CML on Bitbucket, I ran into an issue because I generated the `REPO_TOKEN` incorrectly. I made a mistake while following `echo -n $USERNAME:$PASSWORD | base64`; left in the first `$`.

`$USERNAME` and `$PASSWORD` are not necessarily variables that are set when the user runs this command. Worst case, the local `$USERNAME` is different from the Bitbucket username.

I am not sure if my proposed solution is the ideal one, but earlier in this page we use `{}` to denote that the user should specify a variable here. Maybe we could use that approach here too.